Preparation for the EditContext adoption.
Introduces AbstractEditContext and moves the text area code into editContext/textArea.
