+++ src/vs/editor/browser/controller/editContext/textArea/textAreaEditContext.css
