+++ src/vs/editor/browser/controller/editContext/textArea/textAreaEditContextState.ts
@@ -3,10 +3,10 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import * as strings from '../../../base/common/strings.js';
-import { Position } from '../../common/core/position.js';
-import { Range } from '../../common/core/range.js';
-import { EndOfLinePreference } from '../../common/model.js';
+import { commonPrefixLength, commonSuffixLength } from '../../../../../base/common/strings.js';
+import { Position } from '../../../../common/core/position.js';
+import { Range } from '../../../../common/core/range.js';
+import { ScreenReaderContentState } from '../screenReaderUtils.js';
 
 export const _debugComposition = false;
 
@@ -17,14 +17,6 @@ export interface ITextAreaWrapper {
 	getSelectionStart(): number;
 	getSelectionEnd(): number;
 	setSelectionRange(reason: string, selectionStart: number, selectionEnd: number): void;
-}
-
-export interface ISimpleModel {
-	getLineCount(): number;
-	getLineMaxColumn(lineNumber: number): number;
-	getValueInRange(range: Range, eol: EndOfLinePreference): string;
-	getValueLengthInRange(range: Range, eol: EndOfLinePreference): number;
-	modifyPosition(position: Position, offset: number): Position;
 }
 
 export interface ITypeData {
@@ -130,12 +122,12 @@ export class TextAreaState {
 		}
 
 		const prefixLength = Math.min(
-			strings.commonPrefixLength(previousState.value, currentState.value),
+			commonPrefixLength(previousState.value, currentState.value),
 			previousState.selectionStart,
 			currentState.selectionStart
 		);
 		const suffixLength = Math.min(
-			strings.commonSuffixLength(previousState.value, currentState.value),
+			commonSuffixLength(previousState.value, currentState.value),
 			previousState.value.length - previousState.selectionEnd,
 			currentState.value.length - currentState.selectionEnd
 		);
@@ -202,8 +194,8 @@ export class TextAreaState {
 			};
 		}
 
-		const prefixLength = Math.min(strings.commonPrefixLength(previousState.value, currentState.value), previousState.selectionEnd);
-		const suffixLength = Math.min(strings.commonSuffixLength(previousState.value, currentState.value), previousState.value.length - previousState.selectionEnd);
+		const prefixLength = Math.min(commonPrefixLength(previousState.value, currentState.value), previousState.selectionEnd);
+		const suffixLength = Math.min(commonSuffixLength(previousState.value, currentState.value), previousState.value.length - previousState.selectionEnd);
 		const previousValue = previousState.value.substring(prefixLength, previousState.value.length - suffixLength);
 		const currentValue = currentState.value.substring(prefixLength, currentState.value.length - suffixLength);
 		const previousSelectionStart = previousState.selectionStart - prefixLength;
@@ -223,65 +215,14 @@ export class TextAreaState {
 			positionDelta: currentSelectionEnd - currentValue.length
 		};
 	}
-}
 
-export class PagedScreenReaderStrategy {
-	private static _getPageOfLine(lineNumber: number, linesPerPage: number): number {
-		return Math.floor((lineNumber - 1) / linesPerPage);
-	}
-
-	private static _getRangeForPage(page: number, linesPerPage: number): Range {
-		const offset = page * linesPerPage;
-		const startLineNumber = offset + 1;
-		const endLineNumber = offset + linesPerPage;
-		return new Range(startLineNumber, 1, endLineNumber + 1, 1);
-	}
-
-	public static fromEditorSelection(model: ISimpleModel, selection: Range, linesPerPage: number, trimLongText: boolean): TextAreaState {
-		// Chromium handles very poorly text even of a few thousand chars
-		// Cut text to avoid stalling the entire UI
-		const LIMIT_CHARS = 500;
-
-		const selectionStartPage = PagedScreenReaderStrategy._getPageOfLine(selection.startLineNumber, linesPerPage);
-		const selectionStartPageRange = PagedScreenReaderStrategy._getRangeForPage(selectionStartPage, linesPerPage);
-
-		const selectionEndPage = PagedScreenReaderStrategy._getPageOfLine(selection.endLineNumber, linesPerPage);
-		const selectionEndPageRange = PagedScreenReaderStrategy._getRangeForPage(selectionEndPage, linesPerPage);
-
-		let pretextRange = selectionStartPageRange.intersectRanges(new Range(1, 1, selection.startLineNumber, selection.startColumn))!;
-		if (trimLongText && model.getValueLengthInRange(pretextRange, EndOfLinePreference.LF) > LIMIT_CHARS) {
-			const pretextStart = model.modifyPosition(pretextRange.getEndPosition(), -LIMIT_CHARS);
-			pretextRange = Range.fromPositions(pretextStart, pretextRange.getEndPosition());
-		}
-		const pretext = model.getValueInRange(pretextRange, EndOfLinePreference.LF);
-
-		const lastLine = model.getLineCount();
-		const lastLineMaxColumn = model.getLineMaxColumn(lastLine);
-		let posttextRange = selectionEndPageRange.intersectRanges(new Range(selection.endLineNumber, selection.endColumn, lastLine, lastLineMaxColumn))!;
-		if (trimLongText && model.getValueLengthInRange(posttextRange, EndOfLinePreference.LF) > LIMIT_CHARS) {
-			const posttextEnd = model.modifyPosition(posttextRange.getStartPosition(), LIMIT_CHARS);
-			posttextRange = Range.fromPositions(posttextRange.getStartPosition(), posttextEnd);
-		}
-		const posttext = model.getValueInRange(posttextRange, EndOfLinePreference.LF);
-
-
-		let text: string;
-		if (selectionStartPage === selectionEndPage || selectionStartPage + 1 === selectionEndPage) {
-			// take full selection
-			text = model.getValueInRange(selection, EndOfLinePreference.LF);
-		} else {
-			const selectionRange1 = selectionStartPageRange.intersectRanges(selection)!;
-			const selectionRange2 = selectionEndPageRange.intersectRanges(selection)!;
-			text = (
-				model.getValueInRange(selectionRange1, EndOfLinePreference.LF)
-				+ String.fromCharCode(8230)
-				+ model.getValueInRange(selectionRange2, EndOfLinePreference.LF)
-			);
-		}
-		if (trimLongText && text.length > 2 * LIMIT_CHARS) {
-			text = text.substring(0, LIMIT_CHARS) + String.fromCharCode(8230) + text.substring(text.length - LIMIT_CHARS, text.length);
-		}
-
-		return new TextAreaState(pretext + text + posttext, pretext.length, pretext.length + text.length, selection, pretextRange.endLineNumber - pretextRange.startLineNumber);
+	public static fromScreenReaderContentState(screenReaderContentState: ScreenReaderContentState) {
+		return new TextAreaState(
+			screenReaderContentState.value,
+			screenReaderContentState.selectionStart,
+			screenReaderContentState.selectionEnd,
+			screenReaderContentState.selection,
+			screenReaderContentState.newlineCountBeforeSelection
+		);
 	}
 }
